+++ src/main/java/com/lxp/domain/user/enums/ActiveStatus.java
@@ -1,4 +1,4 @@
-package com.lxp.domain.user;
+package com.lxp.domain.user.enums;
 
 public enum ActiveStatus {
     ACTIVE,
