+++ src/main/java/com/lxp/domain/user/enums/UserRole.java
@@ -1,4 +1,4 @@
-package com.lxp.domain.user;
+package com.lxp.domain.user.enums;
 
 public enum UserRole {
     STUDENT,
