+++ src/main/java/com/lxp/util/QueryType.java
@@ -3,7 +3,8 @@ package com.lxp.util;
 public enum QueryType {
 
     // user_queries.xml
-    USER_FIND_BY_ID("user.findById");
+    USER_FIND_BY_ID("user.findById"),
+    USER_FIND_BY_NAME("user.findByName");
 
     private final String key;
